Refactor ProfileInfo component to new User directory and remove old implementation move another directory profileinfo jsx, and profilestats
